# dist/public/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "Partial_Token_": {
                "properties": {},
                "type": "object",
                "description": "Make all properties in T optional"
            },
...
            "ILoginInput": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "IRefreshTokenInput": {
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    }
                },
                "required": [
                    "refreshToken"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "IAcceptInviteInput": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "action": {
                        "type": "string",
                        "enum": [
                            "accept",
                            "reject"
                        ]
                    }
                },
                "required": [
                    "token"
                ],
                "type": "object",
                "additionalProperties": false
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/auth/login": {
            "post": {
                "operationId": "Login",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Partial_Token_"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ILoginInput"
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/logout": {
            "get": {
                "operationId": "Logout",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": []
            }
        },
...
        "/auth/refresh-token": {
            "post": {
                "operationId": "RefreshToken",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Partial_Token_"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IRefreshTokenInput"
                            }
                        }
                    }
                }
            }
        },
...
        "/admin-auth/login": {
            "post": {
                "operationId": "Login",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Partial_Token_"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Admin Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ILoginInput"
                            }
                        }
                    }
                }
            }
        },
        "/admin-auth/logout": {
            "get": {
                "operationId": "Logout",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Admin Auth"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": []
            }
        },
        "/admin-auth/refresh-token": {
            "post": {
                "operationId": "RefreshToken",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Partial_Token_"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Admin Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IRefreshTokenInput"
                            }
                        }
                    }
                }
            }
        },
        "/admin-auth/accept-invite": {
            "post": {
                "operationId": "AcceptInvite",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Admin Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IAcceptInviteInput"
                            }
                        }
                    }
                }
            }
        }
    },
    "info": {
        "license": {
            "name": "ISC"
        },
        "contact": {},
        "title": "MD Exchange API",
        "description": "API for managing MD Exchange operations",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "localhost dev"
        }
    ]
}